# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Users Api",
    "description": "Users Api",
    "version": "1.0.0"
  },
  "host": "cse341-crud-uv92.onrender.com",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "securityDefinitions": {
    "oauth2": {
      "type": "oauth2",
      "flow": "implicit",
      "authorizationUrl": "https://github.com/login/oauth/authorize",
      "scopes": {
        "user": "Acceso a la información del usuario"
      }
    }
  },
  "paths": {
    "/": {
      "get": {
        "tags": ["Hello World"],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/login": {
      "get": {
        "tags": ["Auth"],
        "summary": "Iniciar sesión con GitHub",
        "description": "Redirige a GitHub para autenticación OAuth",
        "security": [{ "oauth2": [] }],
        "responses": {
          "302": {
            "description": "Redirección a GitHub"
          }
        }
      }
    },
    "/auth/github/callback": {
      "get": {
        "tags": ["Auth"],
        "summary": "Callback de GitHub",
        "description": "GitHub redirige aquí después del login",
        "responses": {
          "302": {
            "description": "Redirección a la página principal"
          }
        }
      }
    },
    "/logout": {
      "get": {
        "tags": ["Auth"],
        "summary": "Cerrar sesión",
        "description": "Cierra la sesión del usuario autenticado",
        "responses": {
          "200": {
            "description": "Sesión cerrada correctamente"
          }
        }
      }
    },
    "/users/": {
      "get": {
        "tags": ["Users"],
        "description": "Obtener lista de usuarios",
        "produces": ["application/json"],
        "security": [{ "oauth2": [] }],
        "responses": {
          "200": { "description": "OK" },
          "401": { "description": "Unauthorized" },
          "500": { "description": "Internal Server Error" }
        }
      },
      "post": {
        "tags": ["Users"],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "firstName": { "example": "any" },
                "lastName": { "example": "any" },
                "email": { "example": "any" },
                "favoriteColor": { "example": "any" },
                "birthday": { "example": "any" }
              }
            }
          }
        ],
        "responses": {
          "201": { "description": "Created" },
          "400": { "description": "Bad Request" },
          "500": { "description": "Internal Server Error" }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": ["Users"],
        "description": "",
        "produces": ["application/json"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "type": "string" }
        ],
        "security": [{ "oauth2": [] }],
        "responses": {
          "200": { "description": "OK" },
          "400": { "description": "Bad Request" },
          "404": { "description": "Not Found" },
          "500": { "description": "Internal Server Error" }
        }
      },
      "put": {
        "tags": ["Users"],
        "description": "",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "type": "string" },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "userName": { "example": "any" },
                "email": { "example": "any" },
                "name": { "example": "any" },
                "ipaddress": { "example": "any" }
              }
            }
          }
        ],
        "responses": {
          "200": { "description": "OK" },
          "400": { "description": "Bad Request" },
          "500": { "description": "Internal Server Error" }
        }
      },
      "delete": {
        "tags": ["Users"],
        "description": "",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "type": "string" }
        ],
        "responses": {
          "200": { "description": "OK" },
          "400": { "description": "Bad Request" },
          "404": { "description": "Not Found" },
          "500": { "description": "Internal Server Error" }
        }
      }
    }
  }
}